chore: refactor handlers in test server
