+++ test-server/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod handle;
 pub mod responses;
 
 /// Returns `main.dummy`
