+++ test-server/src/main.rs
@@ -1,18 +1,11 @@
 use anyhow::Result;
-use log::{error, info};
-use lsp_server::{Connection, Message, Notification, Request, RequestId, Response};
+use log::info;
+use lsp_server::{Connection, Message};
 use lsp_types::{
-    notification::{DidOpenTextDocument, Notification as _, PublishDiagnostics},
-    request::{
-        DocumentDiagnosticRequest, Formatting, GotoDefinition, References, Rename, Request as _,
-    },
     DiagnosticOptions, DiagnosticServerCapabilities, InitializeParams, OneOf, ServerCapabilities,
-    TextDocumentSyncCapability, TextDocumentSyncKind, Uri, WorkDoneProgressOptions,
+    TextDocumentSyncCapability, TextDocumentSyncKind, WorkDoneProgressOptions,
 };
-use test_server::responses::{
-    get_definition_response, get_diagnostics_response, get_formatting_response,
-    get_references_response, get_rename_response,
-};
+use test_server::handle::{handle_notification, handle_request};
 
 /// Entry point of the lsp server. Connects to the client and enters the main loop
 ///
@@ -92,189 +85,4 @@ fn main_loop(connection: &Connection) -> Result<()> {
         }
     }
     Ok(())
-}
-
-fn cast_req<R>(req: Request) -> Result<(RequestId, R::Params)>
-where
-    R: lsp_types::request::Request,
-    R::Params: serde::de::DeserializeOwned,
-{
-    match req.extract(R::METHOD) {
-        Ok(value) => Ok(value),
-        Err(e) => Err(anyhow::anyhow!("Error: {e}")),
-    }
-}
-
-fn cast_notif<R>(notif: Notification) -> Result<R::Params>
-where
-    R: lsp_types::notification::Notification,
-    R::Params: serde::de::DeserializeOwned,
-{
-    match notif.extract(R::METHOD) {
-        Ok(value) => Ok(value),
-        Err(e) => Err(anyhow::anyhow!("Error: {e}")),
-    }
-}
-
-/// Handles `Request`s from the lsp client.
-///
-/// By convention, the `response_num` value specifying which pre-determined response
-/// to send back is taken from the first line number in `params`, if available.
-/// Data passed via other means will be specified in a comment
-///
-/// # Errors
-///
-/// Returns errors from any of the handler functions. The majority of error sources
-/// are failures to send a response via `connection`
-///
-/// # Panics
-///
-/// Panics if JSON encoding of a response fails or if a json request fails to cast
-/// into its equivalent in memory struct
-#[allow(clippy::too_many_lines)]
-pub fn handle_request(req: Request, connection: &Connection) -> Result<()> {
-    match req.method.as_str() {
-        References::METHOD => {
-            let (id, params) =
-                cast_req::<References>(req).expect("Failed to cast References request");
-            info!(
-                "Received `{}` request ({id}): {params:?}",
-                References::METHOD
-            );
-            let response_num = params.text_document_position.position.line;
-            info!("response_num: {response_num}");
-            let Some(resp) = get_references_response(response_num) else {
-                error!("Invalid response number: {response_num}");
-                return Ok(());
-            };
-            let result = serde_json::to_value(&resp).unwrap();
-
-            let result = Response {
-                id,
-                result: Some(result),
-                error: None,
-            };
-            return Ok(connection.sender.send(Message::Response(result))?);
-        }
-        Formatting::METHOD => {
-            let (id, params) =
-                cast_req::<Formatting>(req).expect("Failed to cast Formatting request");
-            info!(
-                "Received `{}` request ({id}): {params:?}",
-                Formatting::METHOD
-            );
-            // `response_num` passed via `params.options.tab_size`
-            let response_num = params.options.tab_size;
-            info!("response_num: {response_num}");
-            let resp = get_formatting_response(response_num).map_or_else(
-                || {
-                    // In this case, we wish to test `FormattingResponse::EndState`
-                    // Send a  reply with no edits (so the start and end state of
-                    // the file matches) to the client so it knows we got the request
-                    // and proceeds with the comparison
-                    info!("Sending response for `FormattingResponse::EndState`");
-                    Vec::new()
-                },
-                |resp| {
-                    info!("Sending response for `FormattingResponse::Response`");
-                    resp
-                },
-            );
-            let result = serde_json::to_value(&resp).unwrap();
-
-            let result = Response {
-                id,
-                result: Some(result),
-                error: None,
-            };
-            return Ok(connection.sender.send(Message::Response(result))?);
-        }
-        Rename::METHOD => {
-            let (id, params) = cast_req::<Rename>(req).expect("Failed to cast Rename request");
-            info!("Received `{}` request ({id}): {params:?}", Rename::METHOD);
-            // `response_num` passed via `params.new_name`
-            let Ok(response_num) = params.new_name.parse() else {
-                error!(
-                    "Failed to parse `new_name` as `response_num`: {}",
-                    params.new_name
-                );
-                return Ok(());
-            };
-            info!("response_num: {response_num}");
-            let Some(resp) = get_rename_response(response_num) else {
-                error!("Invalid response number: {response_num}");
-                return Ok(());
-            };
-            let result = serde_json::to_value(&resp).unwrap();
-
-            let result = Response {
-                id,
-                result: Some(result),
-                error: None,
-            };
-            return Ok(connection.sender.send(Message::Response(result))?);
-        }
-        GotoDefinition::METHOD => {
-            let (id, params) =
-                cast_req::<GotoDefinition>(req).expect("Failed to cast GotoDefinition request");
-            info!(
-                "Received `{}` request ({id}): {params:?}",
-                GotoDefinition::METHOD
-            );
-            let response_num = params.text_document_position_params.position.line;
-            info!("response_num: {response_num}");
-            let Some(resp) = get_definition_response(response_num) else {
-                error!("Invalid response number: {response_num}");
-                return Ok(());
-            };
-            let result = serde_json::to_value(&resp).unwrap();
-
-            let result = Response {
-                id,
-                result: Some(result),
-                error: None,
-            };
-            return Ok(connection.sender.send(Message::Response(result))?);
-        }
-        DocumentDiagnosticRequest::METHOD => {
-            let (id, params) = cast_req::<DocumentDiagnosticRequest>(req)
-                .expect("Failed to cast DocumentDiagnosticRequest request");
-            info!(
-                "Received `{}` request ({id}): {params:?}",
-                DocumentDiagnosticRequest::METHOD
-            );
-            send_diagnostic_resp(&params.text_document.uri, connection)?;
-        }
-        method => error!("Unimplemented request format: {method:?}\n{req:?}"),
-    }
-
-    Ok(())
-}
-
-fn handle_notification(notif: Notification, connection: &Connection) -> Result<()> {
-    match notif.method.as_str() {
-        DidOpenTextDocument::METHOD => {
-            let did_open_params = cast_notif::<DidOpenTextDocument>(notif)?;
-            info!(
-                "Received `{}` notification: {did_open_params:?}",
-                DidOpenTextDocument::METHOD
-            );
-            send_diagnostic_resp(&did_open_params.text_document.uri, connection)?;
-        }
-        method => error!("Unimplemented notification format: {method:?}\n{notif:?}"),
-    }
-    Ok(())
-}
-
-// TODO: Move this once we refactor
-fn send_diagnostic_resp(uri: &Uri, connection: &Connection) -> Result<()> {
-    let publish_params = get_diagnostics_response(uri);
-    let result = serde_json::to_value(&publish_params).unwrap();
-
-    let notif = Notification {
-        method: PublishDiagnostics::METHOD.to_string(),
-        params: result,
-    };
-
-    Ok(connection.sender.send(Message::Notification(notif))?)
 }
